[REFACTOR] divide Auth Section and Update MyPage Style
인증 영역 분할 및 스타일 수정
